### 2024-09-21 #7261 @editorialbot
[PRE REVIEW]: AutoPDEx: An Automized Partial Differential Equation solver based on JAX
**Submitting author:** @BodeTobias (<a href="http://orcid.org/0000-0001-7941-7658">Tobias Bode</a>)
**Repository:** https://github.com/BodeTobias/AutoPDEx
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.0
