### 2024-09-24 #7261 @BodeTobias
Dear @danielskatz,
thank you for being the editor for my submission! I think the authors and reviewers of the first two similar JOSS papers could be a good fit. The first two of the following list I also found in the list of reviewers:

mrava87
celliern
poulson
dawid-zwicker
mstimberg
