### 2024-09-25 #7261 @danielskatz
👋 @Sm00thix, @parmentelat, @basileMarchand - Would any of you be willing to review this JOSS submission?
