### 2024-09-25 #7261 @danielskatz
👋 @jeremiecoullon, @canyon289, @ColCarroll - Would any of you be willing to review this JOSS submission?
