### 2024-09-26 #7261 @jeremiecoullon
Hello! I don't know a lot about PDE solvers, and I have a high workflow at the moment so I can't help with this unfortunately. :(
