### 2024-09-27 #7261 @parmentelat
hello @danielskatz
thanks for the invitation, but I have very little to do with pde's so this is definitely very far from my comfort area, and I will have to decline, but thanks again
