### 2024-09-27 #7300 @editorialbot
[REVIEW]: AutoPDEx: An Automized Partial Differential Equation solver based on JAX
**Submitting author:** @BodeTobias (<a href="http://orcid.org/0000-0001-7941-7658">Tobias Bode</a>)
**Repository:** https://github.com/BodeTobias/AutoPDEx
**Branch with paper.md** (empty if default branch):
**Version:** v1.1.4
