### 2024-09-27 #7300 @danielskatz
@openjournals/dev - Any idea what's going on with generating the PDF here? This looks like some type of system error
