### 2024-09-27 #7300 @danielskatz
(and sorry the PDF isn't immediately available - we'll figure this out and get it made ASAP - you can refer to the version from the pre-review issue for now: https://github.com/openjournals/joss-papers/blob/joss.07261/joss.07261/10.21105.joss.07261.pdf)
