### 2024-09-28 BodeTobias/AutoPDEx#1 @celliern
Pytest is mandatory to run the library
Hello! It seems that you included a `pytest` import at your [root `__init__.py` file](https://github.com/BodeTobias/AutoPDEx/blob/04ecb9456fe4f426f10178c424376965f9faf429/autopdex/__init__.py#L2), without including pytest in your dependencies list.
You should either remove the pytest statement in your `__init__.py` file or add it as a mandatory deps in the `pyproject.toml`.
