### 2024-09-28 #7300 @celliern
A small error make the lib non-importable in an empty env (easily fixable)

https://github.com/BodeTobias/AutoPDEx/issues/1
