### 2024-09-30 3cb5c1a
Fixing DOIs and pytest dependency
