+++ paper/paper.bib
@@ -12,7 +12,8 @@
   journal={Advances in neural information processing systems},
   volume={35},
   pages={5230--5242},
-  year={2022}
+  year={2022},
+  doi={https://doi.org/10.48550/arXiv.2105.15183}
 }
 
 @software{flax2020github,
@@ -27,14 +28,16 @@
   title={Equinox: neural networks in JAX via callable PyTrees and filtered transformations},
   author={Kidger, Patrick and Garcia, Cristian},
   journal={arXiv preprint arXiv:2111.00254},
-  year={2021}
+  year={2021},
+  doi={https://doi.org/10.48550/arXiv.2111.00254}
 }
 
 @article{balay2019petsc,
   title={PETSc users manual},
   author={Balay, Satish and Abhyankar, Shrirang and Adams, Mark and Brown, Jed and Brune, Peter and Buschelman, Kris and Dalcin, Lisandro and Dener, Alp and Eijkhout, Victor and Gropp, William and others},
   year={2019},
-  publisher={Argonne National Laboratory}
+  publisher={Argonne National Laboratory},
+  doi={https://doi.org/10.2172/1577437}
 }
 
 @article{schenk2004solving,
@@ -45,7 +48,8 @@
   number={3},
   pages={475--487},
   year={2004},
-  publisher={Elsevier}
+  publisher={Elsevier},
+  doi = {https://doi.org/10.1016/j.future.2003.07.011}
 }
 
 @article{geuzaine2009gmsh,
@@ -56,7 +60,8 @@
   number={11},
   pages={1309--1331},
   year={2009},
-  publisher={Wiley Online Library}
+  publisher={Wiley Online Library},
+  doi={https://doi.org/10.1002/nme.2579}
 }
 
 @article{ahrens200536,
@@ -66,14 +71,16 @@
   volume={717},
   pages={50038--1},
   year={2005},
-  publisher={Citeseer}
+  publisher={Citeseer},
+  doi={https://doi.org/10.1016/B978-012387582-2/50038-1}
 }
 
 @book{korelc2016automation,
   title={Automation of Finite Element Methods},
   author={Korelc, Joze and Wriggers, Peter},
   year={2016},
-  publisher={Springer}
+  publisher={Springer},
+  doi={https://doi.org/10.1007/978-3-319-39005-5}
 }
 
 @article{sullivan2019pyvista,
@@ -84,5 +91,7 @@
   number={37},
   pages={1450},
   year={2019},
-  publisher={The Open Journal}
+  publisher={The Open Journal},
+  doi = {https://doi.org/10.21105/joss.01450}
 }
+
