+++ pyproject.toml
@@ -30,7 +30,8 @@ dependencies = [
     "pyvista>=0.43.3",
     "scipy>=1.13.0",
     "scikit-optimize>=0.9.0",
-    "sparsejac>=0.1.3"
+    "sparsejac>=0.1.3",
+    "pytest>=8.3.2"
 ]
 
 [project.optional-dependencies]
