### 2024-09-30 38347b9
Adding DOIs and pytest dependency
