+++ docs/conf.py
@@ -13,7 +13,7 @@ sys.path.insert(0, os.path.abspath('..'))
 project = 'AutoPDEx'
 copyright = '2024, Tobias Bode'
 author = 'Tobias Bode'
-release = '0.0.2'
+release = '1.0.1'
 
 # -- General configuration ---------------------------------------------------
 # https://www.sphinx-doc.org/en/master/usage/configuration.html#general-configuration
