+++ pyproject.toml
@@ -4,7 +4,7 @@ build-backend = "setuptools.build_meta"
 
 [project]
 name = "autopdex"
-version = "1.0.0"
+version = "1.0.1"
 description = "An automized PDE solver for accelerators based on JAX"
 readme = "README.md"
 authors = [
