### 2024-09-30 BodeTobias/AutoPDEx#1 @BodeTobias
Hello @celliern, thank you for pointing that out! I have added pytest to the dependencies.
