+++ paper/paper.bib
@@ -13,7 +13,7 @@
   volume={35},
   pages={5230--5242},
   year={2022},
-  doi={https://doi.org/10.48550/arXiv.2105.15183}
+  doi={10.48550/arXiv.2105.15183}
 }
 
 @software{flax2020github,
@@ -29,7 +29,7 @@
   author={Kidger, Patrick and Garcia, Cristian},
   journal={arXiv preprint arXiv:2111.00254},
   year={2021},
-  doi={https://doi.org/10.48550/arXiv.2111.00254}
+  doi={10.48550/arXiv.2111.00254}
 }
 
 @article{balay2019petsc,
@@ -37,7 +37,7 @@
   author={Balay, Satish and Abhyankar, Shrirang and Adams, Mark and Brown, Jed and Brune, Peter and Buschelman, Kris and Dalcin, Lisandro and Dener, Alp and Eijkhout, Victor and Gropp, William and others},
   year={2019},
   publisher={Argonne National Laboratory},
-  doi={https://doi.org/10.2172/1577437}
+  doi={10.2172/1577437}
 }
 
 @article{schenk2004solving,
@@ -49,7 +49,7 @@
   pages={475--487},
   year={2004},
   publisher={Elsevier},
-  doi = {https://doi.org/10.1016/j.future.2003.07.011}
+  doi = {10.1016/j.future.2003.07.011}
 }
 
 @article{geuzaine2009gmsh,
@@ -61,7 +61,7 @@
   pages={1309--1331},
   year={2009},
   publisher={Wiley Online Library},
-  doi={https://doi.org/10.1002/nme.2579}
+  doi={10.1002/nme.2579}
 }
 
 @article{ahrens200536,
@@ -72,7 +72,7 @@
   pages={50038--1},
   year={2005},
   publisher={Citeseer},
-  doi={https://doi.org/10.1016/B978-012387582-2/50038-1}
+  doi={10.1016/B978-012387582-2/50038-1}
 }
 
 @book{korelc2016automation,
@@ -80,7 +80,7 @@
   author={Korelc, Joze and Wriggers, Peter},
   year={2016},
   publisher={Springer},
-  doi={https://doi.org/10.1007/978-3-319-39005-5}
+  doi={10.1007/978-3-319-39005-5}
 }
 
 @article{sullivan2019pyvista,
@@ -92,6 +92,6 @@
   pages={1450},
   year={2019},
   publisher={The Open Journal},
-  doi = {https://doi.org/10.21105/joss.01450}
+  doi = {10.21105/joss.01450}
 }
 
