### 2024-10-10 c7f289b
Fixing int aux data in solver
