+++ pyproject.toml
@@ -1,10 +1,9 @@
 [build-system]
-requires = ["setuptools>=61.0", "wheel"]
+requires = ["setuptools>=61.0", "wheel", "setuptools-scm"]
 build-backend = "setuptools.build_meta"
 
 [project]
 name = "autopdex"
-version = "1.0.3"
 description = "An automized PDE solver for accelerators based on JAX"
 readme = "README.md"
 authors = [
