+++ .github/workflows/tests.yml
@@ -26,8 +26,6 @@ jobs:
           sudo apt-get update
           sudo apt-get install -y libglu1-mesa
           python -m pip install --upgrade pip
-          pip install setuptools
-          pip install pytest
           pip install .
 
       - name: Run tests
