+++ pyproject.toml
@@ -50,3 +50,7 @@ where = ["."]
 
 [tool.setuptools]
 include-package-data = true
+
+[tool.setuptools_scm]
+version_scheme = "post-release"
+local_scheme = "no-local-version"
