+++ .github/workflows/publish_to_pypi.yml
@@ -13,7 +13,9 @@ jobs:
       # Step 1: Check out the repository code
       - name: Check out repository
         uses: actions/checkout@v3
-
+        with:
+          fetch-depth: 0
+    
       # Step 2: Set up Python environment
       - name: Set up Python
         uses: actions/setup-python@v4
