+++ .github/workflows/tests.yml
@@ -15,6 +15,8 @@ jobs:
     steps:
       - name: Check out the repository
         uses: actions/checkout@v3
+        with:
+          fetch-depth: 0
 
       - name: Set up Python
         uses: actions/setup-python@v4
@@ -26,6 +28,7 @@ jobs:
           sudo apt-get update
           sudo apt-get install -y libglu1-mesa
           python -m pip install --upgrade pip
+          pip install setuptools-scm
           pip install .
 
       - name: Run tests
