+++ docs/conf.py
@@ -8,11 +8,14 @@
 
 import os
 import sys
+import setuptools_scm
 sys.path.insert(0, os.path.abspath('..'))
 
 project = 'AutoPDEx'
 copyright = '2024, Tobias Bode'
 author = 'Tobias Bode'
+# version = setuptools_scm.get_version()
+version = '1.0.3'
 
 # -- General configuration ---------------------------------------------------
 # https://www.sphinx-doc.org/en/master/usage/configuration.html#general-configuration
@@ -57,6 +60,20 @@ html_theme = 'sphinx_rtd_theme'
 html_logo = '_static/logo.png'
 html_favicon = ''
 html_static_path = ['_static']
+html_theme_options = {
+    'display_version': True,
+    'prev_next_buttons_location': 'bottom',
+    'style_external_links': True,
+}
+
+# Add the link to the source code
+html_context = {
+    'display_github': True,  # Integrates with GitHub
+    'github_user': 'BodeTobias',  # Your GitHub username
+    'github_repo': 'AutoPDEx',  # Your repository name
+    'github_version': 'main',  # The version of the repository (branch name)
+    'conf_py_path': '/docs/',  # Path to your docs directory
+}
 
 # Custom CSS
 def setup(app):
