### 2024-10-18 #7300 @celliern
Hello ! I plan the finish the review this week-end. Meanwhile, some element may be improved in the documentation :

- a specific community guideline page may be welcome, maybe with a code of conduct and the tooling used / required (syntax convention, test suite needed, etc...)
- the paper lack of a state of the field : maybe there is no alternative using dag-like computation? Even if, mentioning other python FEM tools (like the fenics project?) and what they are missing compared to autoPDEx (like autograd) could be nice
- the statement of need in the documentation could be extended a bit
- if possible, having a proper readme in the repository (with a quick introduction + installation instruction, and maybe a really short example) would make the repository more welcoming to new users

I will do some example reproducibility this week-end
