### 2024-10-21 1fe4035
Integrating DOFs as dicts for mixed problems
