+++ autopdex/implicit_diff.py
@@ -43,6 +43,10 @@ import jax
 import jax.numpy as jnp
 import numpy as np
 
+from autopdex import utility
+
+# Fixme: test with dict dofs
+
 def tree_scalar_mul(scalar, tree_x):
   """Compute scalar * tree_x."""
   return jax.tree_util.tree_map(lambda x: scalar * x, tree_x)
@@ -124,6 +128,8 @@ def _root_vjp(residual_fun: Callable,
     each argument. Each ``vjps[i]`` has the same pytree structure as
     ``args[i]``.
   """
+  if free_dofs is not None:
+    free_dofs_flat = utility.dict_flatten(free_dofs)
 
   @jax.custom_vjp
   def linear_solver_fun_vjp(A, b):
@@ -143,11 +149,9 @@ def _root_vjp(residual_fun: Callable,
           result_dot = linear_solver_fun_vjp(A.T, g)
 
           size = A.shape[0]
-          idx = jnp.arange(size)[free_dofs]
-          result_tmp = jnp.zeros((size,), dtype=jnp.float_)
-          result_tmp = result_tmp.at[idx].set(result)
-          result_dot_tmp = jnp.zeros((size,), dtype=jnp.float_)
-          result_dot_tmp = result_dot_tmp.at[idx].set(result_dot)
+          empty = jnp.zeros((size,), dtype=jnp.float_)
+          result_tmp = utility.mask_op(empty, free_dofs_flat, result, 'set')
+          result_dot_tmp = utility.mask_op(empty, free_dofs_flat, result_dot, 'set')
 
           # # Use the sparsity structure of A to build Fx_dot
           indices = jnp.asarray(A.indices, dtype=jnp.int64)
@@ -182,41 +186,33 @@ def _root_vjp(residual_fun: Callable,
   v = tree_scalar_mul(-1, cotangent)
 
   if free_dofs is not None:
-    constrained_dofs = np.invert(free_dofs)
-    idx_imposed = jnp.arange(sol.flatten().shape[0])[constrained_dofs]
+    constrained_dofs = utility.reshape_as(np.invert(utility.dict_flatten(free_dofs)), free_dofs)
 
-    v_flat = v.flatten()[free_dofs]
-    u = jnp.zeros_like(v)
-    u_flat = u.flatten()
-    idx_free = jnp.arange(u_flat.shape[0])[free_dofs]
+    v_flat = utility.dict_flatten(v)[free_dofs_flat]
     u_f = diffable_solve_fun(mat.T, v_flat)
-    u_flat = u_flat.at[idx_free].set(u_f)
+    u_flat = utility.mask_op(utility.dict_flatten(utility.dict_zeros_like(v)), free_dofs_flat, u_f, 'set')
    
     def fun_args(*args):
       def residual_fun_tmp(sol, *args):
-        dirichlet_values = args[0]['dirichlet conditions'].flatten()[constrained_dofs]
-        sol_with_bc = sol.flatten()
-        sol_with_bc = sol_with_bc.at[idx_imposed].set(dirichlet_values)
-        return residual_fun(sol_with_bc.reshape(sol.shape), *args)
+        dirichlet_values = args[0]['dirichlet conditions']
+        sol_with_bc = utility.mask_op(sol, constrained_dofs, dirichlet_values, 'set')
+        return residual_fun(sol_with_bc, *args)
       return residual_fun_tmp(sol, *args) 
 
-    u = u_flat.reshape(v.shape)
+    u = utility.reshape_as(u_flat, v)
   else:
     def fun_args(*args):
       return residual_fun(sol, *args)
 
-    v_flat = v.flatten()
-    u = diffable_solve_fun(mat.T, v_flat).reshape(v.shape)
+    v_flat = utility.dict_flatten(v)
[…]
