### 2024-10-21 #7300 @danielskatz
👋 @BodeTobias - please do work on the issues mentioned [above](https://github.com/openjournals/joss-reviews/issues/7300#issuecomment-2422437595) by @celliern
