+++ paper/paper.bib
@@ -95,3 +95,12 @@
   doi = {10.21105/joss.01450}
 }
 
+@article{alnaes2015fenics,
+  title={The FEniCS project version 1.5},
+  author={Aln{\ae}s, Martin and Blechta, Jan and Hake, Johan and Johansson, August and Kehlet, Benjamin and Logg, Anders and Richardson, Chris and Ring, Johannes and Rognes, Marie E and Wells, Garth N},
+  journal={Archive of numerical software},
+  volume={3},
+  number={100},
+  year={2015},
+  doi = {10.11588/ans.2015.100.20553}
+}
