+++ autopdex/utility.py
@@ -30,10 +30,13 @@ from jax.experimental import sparse
 
 
 def jit_with_docstring(static_argnames=None):
+    """JIT wrapper that keeps the originial docstring of the function."""
     def decorator(fun):
         # Define the jitted function with optional static_argnames
         if static_argnames is not None:
-            jitted_fun = jax.tree_util.Partial(jax.jit, static_argnames=static_argnames)(fun)
+            jitted_fun = jax.tree_util.Partial(
+                jax.jit, static_argnames=static_argnames
+            )(fun)
         else:
             jitted_fun = jax.jit(fun)
 
@@ -43,18 +46,21 @@ def jit_with_docstring(static_argnames=None):
             return jitted_fun(*args, **kwargs)
 
         return wrapper
+
     return decorator
 
+
 def dict_zeros_like(arr, **keyargs):
-  """Wrapper around zeros_like, that works also for dicts with jnp.ndarray entries."""
-  if isinstance(arr, dict):
-    res = {}
-    fields = list(arr.keys())
-    for key in fields:
-      res[key] = jnp.zeros_like(arr[key], **keyargs)
-    return res
-  else: 
-    return jnp.zeros_like(arr, **keyargs)
+    """Wrapper around zeros_like, that works also for dicts with jnp.ndarray entries."""
+    if isinstance(arr, dict):
+        res = {}
+        fields = list(arr.keys())
+        for key in fields:
+            res[key] = jnp.zeros_like(arr[key], **keyargs)
+        return res
+    else:
+        return jnp.zeros_like(arr, **keyargs)
+
 
 def dict_flatten(arr):
     """
@@ -76,11 +82,16 @@ def dict_flatten(arr):
         if not flat_arrays:
             return np.array([]) if isinstance(flat_array, np.ndarray) else jnp.array([])
         else:
-            return jnp.concatenate(flat_arrays) if isinstance(flat_arrays[0], jnp.ndarray) else np.concatenate(flat_arrays)
+            return (
+                jnp.concatenate(flat_arrays)
+                if isinstance(flat_arrays[0], jnp.ndarray)
+                else np.concatenate(flat_arrays)
+            )
     elif isinstance(arr, np.ndarray) or isinstance(arr, jnp.ndarray):
         return arr.flatten()
     else:
         raise TypeError("Input must be a nested dict of arrays or an array.")
+
 
 def reshape_as(flat_array, signature_array):
     """
@@ -99,7 +110,11 @@ def reshape_as(flat_array, signature_array):
         for key in sorted(signature_array.keys()):
             sig_value = signature_array[key]
             # Calculate the number of elements needed for this array
-            size = sig_value.size if isinstance(sig_value, (np.ndarray, jnp.ndarray)) else 0
+            size = (
+                sig_value.size
+                if isinstance(sig_value, (np.ndarray, jnp.ndarray))
+                else 0
+            )
             end = start + size
             # Extract the relevant slice from flat_array
             slice_flat = flat_array[start:end]
@@ -111,9 +126,13 @@ def reshape_as(flat_array, signature_array):
                 # Reshape the slice to match the shape of the signature array
                 reshaped_array = slice_flat.reshape(sig_value.shape)
                 # Convert to the same type as sig_value if necessary
-                if isinstance(sig_value, np.ndarray) and not isinstance(reshaped_array, np.ndarray):
+                if isinstance(sig_value, np.ndarray) and not isinstance(
+                    reshaped_array, np.ndarray
+                ):
                     reshaped_array = np.array(reshaped_array)
-                elif isinstance(sig_value, jnp.ndarray) and not isinstance(reshaped_array, jnp.ndarray):
[…]
