+++ autopdex/__init__.py
@@ -1,5 +1,6 @@
 # autopdex/__init__.py
 import pytest
 
+
 def run_tests():
     pytest.main(["-v", "tests/"])
