+++ README.md
@@ -135,7 +135,7 @@ meshio.Mesh(
 
 You are warmly invited to contribute to the project. For larger developments, please get in touch beforehand in order to circumvent double work. 
 
-For detailed information on how to contribute, please see our [Contribution Guidelines](https://github.com/BodeTobias/AutoPDEx/CONTRIBUTING.md)
+For detailed information on how to contribute, please see our [Contribution Guidelines](https://github.com/BodeTobias/AutoPDEx/blob/main/CONTRIBUTING.md)
 
 ## License
 
