### 2024-10-22 ade552b
Update pyproject.toml
Signed-off-by: Tobias Bode <149093216+BodeTobias@users.noreply.github.com>
