+++ autopdex/seeder.py
@@ -1833,9 +1833,6 @@ def int_pts_ref_tri(order):
                 ),
             )
         case 3:
-            # return (jnp.asarray([[0.333333333333333,0.2,0.2,0.6],
-            #                      [0.333333333333333,0.6,0.2,0.2]]).transpose(),
-            #                      (1/2)*jnp.asarray([-0.5625,0.520833333333333,0.520833333333333,0.520833333333333]))
             return (
                 jnp.asarray(
                     [
@@ -1993,18 +1990,6 @@ def int_pts_ref_tri(order):
                 ),
             )
         case 7:
-            # return (jnp.asarray([[0.333333333333333,
-            #         0.065130102902216,0.065130102902216,0.869739794195568,0.048690315425316,
-            #         0.312865496004874,0.638444188569810,0.312865496004874,0.048690315425316,
-            #         0.638444188569810,0.260345966079040,0.260345966079040,0.479308067841920],
-            #         [0.333333333333333,0.869739794195568,0.065130102902216,0.065130102902216,
-            #         0.638444188569810,0.048690315425316,0.312865496004874,0.638444188569810,
-            #         0.312865496004874,0.048690315425316,0.479308067841920,0.260345966079040,
-            #         0.260345966079040]]).transpose(),
-            #         (1/2)*jnp.asarray([-0.149570044467682,0.053347235608838,0.053347235608838,
-            #         0.053347235608838,0.077113760890257,0.077113760890257,0.077113760890257,
-            #         0.077113760890257,0.077113760890257,0.077113760890257,0.175615257433208,
-            #         0.175615257433208,0.175615257433208]))
             return (
                 jnp.asarray(
                     [
