### 2024-11-24 #7300 @danielskatz
@BodeTobias - have you been able to address the [issues](https://github.com/openjournals/joss-reviews/issues/7300#issuecomment-2422437595) @celliern brought up?
