+++ paper/paper.bib
@@ -104,3 +104,15 @@
   year={2015},
   doi = {10.11588/ans.2015.100.20553}
 }
+
+@article{mitusch2019dolfin,
+  title={dolfin-adjoint 2018.1: automated adjoints for FEniCS and Firedrake},
+  author={Mitusch, Sebastian and Funke, Simon and Dokken, J{\o}rgen},
+  journal={Journal of Open Source Software},
+  volume={4},
+  number={38},
+  pages={1292},
+  year={2019},
+  publisher={The Open Journal},
+  doi = {10.21105/joss.01292}
+}
