+++ paper/paper.md
@@ -28,9 +28,7 @@ The `implicit_diff` module provides a wrapper to make the solution methods diffe
 
 The efficient and accurate solution of partial differential equations (PDEs) is a central task in many scientific and engineering applications. Python has become a popular platform for PDE solving due to its versatility and the availability of mature libraries such as FEniCS [@alnaes2015fenics] with dolfin-adjoint [@mitusch2019dolfin]. However, many existing tools rely on traditional numerical approaches and require manual implementation of residuals and tangents, or use domain-specific symbolic manipulations.
 
-AutoPDEx uses the code-to-code transformation library JAX to compute local derivatives such as spatial gradients, as well as residual and tangent contributions, through automatic differentiation. By leveraging automatic implicit differentiation, the entire solution procedure can also be differentiated in a unified way. This and the modular structure of AutoPDEx make it attractive for research applications in which sensitivities are advantageous, as e.g. in material parameter identification, topology optimization, uncertainty estimation, and multi-scale analysis [@korelc2016automation]. 
-
-By building on JAX, it further enables the smooth combination of numerical simulations with machine learning models from the JAX ecosystem, as for instance those provided by Flax [@flax2020github] and Equinox [@kidger2021equinox]. This capability allows incorporating data-driven approaches into traditional simulation workflows. 
+AutoPDEx uses the code-to-code transformation library JAX to compute local derivatives such as spatial gradients, as well as residual and tangent contributions, through automatic differentiation. By leveraging automatic implicit differentiation, the entire solution procedure can also be differentiated in a unified way. This and the modular structure of AutoPDEx make it attractive for research applications in which sensitivities are advantageous, as e.g. in material parameter identification, topology optimization, uncertainty estimation, and multi-scale analysis [@korelc2016automation]. By building on JAX, AutoPDEx further enables the smooth combination of numerical simulations with machine learning models from the JAX ecosystem, as for instance those provided by Flax [@flax2020github] and Equinox [@kidger2021equinox]. This capability allows incorporating data-driven approaches into traditional simulation workflows. 
 
 The analysis in AutoPDEx can be used in combination with established tools like Gmsh [@geuzaine2009gmsh] for mesh generation and PyVista [@sullivan2019pyvista] or ParaView [@ahrens200536] for visualization, providing a comprehensive solution from model preparation to analysis.
 
