+++ paper/paper.md
@@ -16,7 +16,7 @@ bibliography: paper.bib
 
 # Summary
 
-`AutoPDEx` is a free and open-source software for solving partial differential equations (PDEs) based on the automatic code transformation capabilities of `JAX` [@jax2018github]. It is designed to provide a modular, flexible, and extendable environment for solving boundary and initial value problems, allowing seamless integration with machine learning algorithms and GPU acceleration through the Accelerated Linear Algebra (XLA) compiler.
+`AutoPDEx` is a free and open-source software for solving partial differential equations (PDEs) based on the automatic code transformation capabilities of `JAX` [@jax2018github]. It is designed to provide a modular, flexible and extendable environment for solving boundary and initial value problems, allowing seamless integration with machine learning algorithms and GPU acceleration through the Accelerated Linear Algebra (XLA) compiler.
 
 At its core, AutoPDEx includes a versatile solver module that supports algorithms such as adaptive load stepping, Newton's method, and nonlinear minimizers. The PDEs to be solved and the chosen variational methods and solution spaces can be specified via user-defined JAX-transformable functions. Pre-built models and ansatz functions are available in the `models` and `spaces` modules. In addition to finite element methods, e.g. mesh-free approaches or neural networks can be used as solution spaces. This flexibility makes AutoPDEx suitable for researchers working at the intersection of numerical analysis and machine learning.
 
@@ -26,9 +26,9 @@ The `implicit_diff` module provides a wrapper to make the solution methods diffe
 
 # Statement of Need
 
-The efficient and accurate solution of partial differential equations (PDEs) is a central task in many scientific and engineering applications. Python has become a popular platform for PDE solving due to its versatility and the availability of mature libraries such as FEniCS [@alnaes2015fenics] with dolfin-adjoint [@mitusch2019dolfin]. However, many existing tools rely on traditional numerical approaches and require manual implementation of residuals and tangents, or use domain-specific symbolic manipulations.
+The efficient and accurate solution of partial differential equations (PDEs) is a central task in many scientific and engineering applications. Python has become a popular platform for PDE solving due to its versatility and the availability of mature libraries such as FEniCS [@alnaes2015fenics] with dolfin-adjoint [@mitusch2019dolfin] for sensitivity analysis. However, many existing tools rely on traditional numerical approaches and require manual implementation of residuals and tangents, or use domain-specific symbolic manipulations.
 
-AutoPDEx uses the code-to-code transformation library JAX to compute local derivatives such as spatial gradients, as well as residual and tangent contributions, through automatic differentiation. By leveraging automatic implicit differentiation, the entire solution procedure can also be differentiated in a unified way. This and the modular structure of AutoPDEx make it attractive for research applications in which sensitivities are advantageous, as e.g. in material parameter identification, topology optimization, uncertainty estimation, and multi-scale analysis [@korelc2016automation]. By building on JAX, AutoPDEx further enables the smooth combination of numerical simulations with machine learning models from the JAX ecosystem, as for instance those provided by Flax [@flax2020github] and Equinox [@kidger2021equinox]. This capability allows incorporating data-driven approaches into traditional simulation workflows. 
[…]
