### 2024-11-27 #7300 @BodeTobias
Dear @celliern,

thank you very much for testing the library and for your suggestions to improve the documentation!

I have created a readme including a short example in the Github repository and added community guidelines. I have also added fenics with dolfin-adjoint to the state of the field and extended the statement of need slightly.

Best regards
Tobias Bode
