### 2024-12-06 b831798
backup
sensitivity analysis working without free_dofs
