+++ autopdex/utility.py
@@ -27,6 +27,7 @@ import jax.numpy as jnp
 import numpy as np
 from flax.core import FrozenDict
 from jax.experimental import sparse
+from equinox import filter_jit
 
 
 def jit_with_docstring(static_argnames=None):
@@ -44,11 +45,9 @@ def jit_with_docstring(static_argnames=None):
         @wraps(fun)
         def wrapper(*args, **kwargs):
             return jitted_fun(*args, **kwargs)
-
         return wrapper
 
     return decorator
-
 
 def dict_zeros_like(arr, **keyargs):
     """Wrapper around zeros_like, that works also for dicts with jnp.ndarray entries."""
@@ -60,6 +59,17 @@ def dict_zeros_like(arr, **keyargs):
         return res
     else:
         return jnp.zeros_like(arr, **keyargs)
+
+def dict_ones_like(arr, **keyargs):
+    """Wrapper around ones_like, that works also for dicts with jnp.ndarray entries."""
+    if isinstance(arr, dict):
+        res = {}
+        fields = list(arr.keys())
+        for key in fields:
+            res[key] = jnp.ones_like(arr[key], **keyargs)
+        return res
+    else:
+        return jnp.ones_like(arr, **keyargs)
 
 
 def dict_flatten(arr):
@@ -73,12 +83,8 @@ def dict_flatten(arr):
     Returns:
         array: A single flat array containing all elements, with the same type as the input arrays.
     """
-    if isinstance(arr, dict):
-        flat_arrays = []
-        # Sort keys for consistent ordering
-        for key in arr.keys():
-            flat_array = dict_flatten(arr[key])
-            flat_arrays.append(flat_array)
+    if isinstance(arr, (dict, FrozenDict)):
+        flat_arrays = [dict_flatten(arr[key]) for key in arr.keys()]
         if not flat_arrays:
             return np.array([]) if isinstance(flat_array, np.ndarray) else jnp.array([])
         else:
@@ -87,11 +93,10 @@ def dict_flatten(arr):
                 if isinstance(flat_arrays[0], jnp.ndarray)
                 else np.concatenate(flat_arrays)
             )
-    elif isinstance(arr, np.ndarray) or isinstance(arr, jnp.ndarray):
+    elif isinstance(arr, (np.ndarray, jnp.ndarray)):
         return arr.flatten()
     else:
         raise TypeError("Input must be a nested dict of arrays or an array.")
-
 
 def reshape_as(flat_array, signature_array):
     """
@@ -164,7 +169,6 @@ def reshape_as(flat_array, signature_array):
     else:
         raise TypeError("signature_array must be an array or a dict of arrays.")
 
-
 def mask_set(array, selection, values):
     """
     Changes values in a JAX array or dict of arrays based on a boolean mask.
@@ -214,17 +218,17 @@ def mask_set(array, selection, values):
             )
         return flat_array.reshape(array.shape)
 
-
 def mask_select(array, selection):
     """
     Selects elements from a JAX array or dictionary of arrays based on a boolean mask.
+    Non-selected elements are set to zero.
 
     Args:
         array (jnp.ndarray or dict): JAX array or dictionary of JAX arrays.
         selection (jnp.ndarray or dict): Boolean mask or dictionary of boolean masks.
 
     Returns:
-        Selected elements as an array or dictionary of arrays.
+        jnp.ndarray or dict: Array or dictionary of arrays with non-selected elements set to zero.
     """
     if isinstance(array, dict):
         # Ensure that 'selection' is also a dictionary with the same keys
@@ -237,14 +241,15 @@ def mask_select(array, selection):
         # Recursively apply 'mask_select' to each element in the dictionary
[…]
