+++ autopdex/solver.py
@@ -33,7 +33,41 @@ import numpy as np
 import scipy as scp
 from flax.core import FrozenDict
 
-from autopdex import assembler, implicit_diff, utility
+from autopdex import assembler, implicit_diff, utility, implicit_diff_v2
+
+
+
+def custom_root_decorator(residual_fun, tangent_fun, lin_solve_fun, free_dofs=None, dirichlet_conditions=None, has_aux=False):
+    def decorator(root_solve_fun):
+        def wrapped(x):
+            # def tangent_solve_fun(g, y):
+            #     J = tangent_fun(x)
+            #     return lin_solve_fun(J, y)
+
+            def tangent_solve_fun(g, y):
+                J = tangent_fun(x)
+                solve = lambda matvec, x: lin_solve_fun(J, x)
+                transpose_solve = lambda vecmat, x: lin_solve_fun(J.T, x)
+                matvec = lambda x: utility.reshape_as(J @ utility.dict_flatten(x), x)
+                return jax.lax.custom_linear_solve(matvec, y, solve, transpose_solve)
+
+            solution = implicit_diff_v2.custom_root(
+                f=residual_fun,
+                initial_guess=x,
+                solve=lambda f_func, x0: root_solve_fun(x0),
+                tangent_solve=tangent_solve_fun,
+                free_dofs=free_dofs,
+                dirichlet_conditions=dirichlet_conditions,
+                has_aux=False
+            )
+            return solution
+            # if has_aux:
+            #     return solution, aux
+            # else:
+            #     return solution
+
+        return wrapped
+    return decorator
 
 
 ### Solvers as specified by the static_settings and settings
@@ -135,7 +169,6 @@ def solver(dofs, settings, static_settings, **kwargs):
         pass
     return sol, infos
 
-
 @utility.jit_with_docstring(
     static_argnames=[
         "static_settings",
@@ -202,25 +235,18 @@ def adaptive_load_stepping(
 
     if implicit_diff_mode is not None:
         # Set-up the decorator for implicit differentiation
-        residual_fun = lambda dofs, settings: assembler.assemble_residual(
-            dofs, settings, static_settings
-        )
-        tangent_fun = lambda dofs, settings: assembler.assemble_tangent(
-            dofs, settings, static_settings
-        )
-        free_dofs = None
-        free_dofs_flat = None
+        residual_fun = lambda dofs, settings: assembler.assemble_residual(dofs, settings, static_settings)
+        tangent_fun = lambda dofs, settings: assembler.assemble_tangent(dofs, settings, static_settings)
+
         try:
-            if isinstance(static_settings["dirichlet dofs"], FrozenDict):
-                free_dofs = {
-                    key: np.invert(np.asarray(static_settings["dirichlet dofs"][key]))
-                    for key in static_settings["dirichlet dofs"].keys()
-                }
-            else:
-                free_dofs = np.invert(np.asarray(static_settings["dirichlet dofs"]))
-            free_dofs_flat = utility.dict_flatten(free_dofs)
+            dirichlet_dofs = settings["dirichlet dofs"]
         except KeyError:
             pass
+        free_dofs = None
+        free_dofs_flat = None
+        if dirichlet_dofs is not None:
+            free_dofs = utility.mask_op(dirichlet_dofs, utility.dict_ones_like(dirichlet_dofs), mode="apply", ufunc=lambda x: ~x)
+            free_dofs_flat = utility.dict_flatten(free_dofs)
 
         solver_backend = static_settings["solver backend"]
         solver_subtype = static_settings["solver"]
@@ -235,7 +261,7 @@ def adaptive_load_stepping(
             case "petsc":
                 n_fields = static_settings["number of fields"]
[…]
