### 2024-12-06 e2903e8
working free_dofs in sensitivities
