+++ autopdex/implicit_diff.py
@@ -41,10 +41,11 @@ from typing import Tuple
 
 import jax
 import jax.numpy as jnp
-import numpy as np
 
 from autopdex import utility
 from flax.core import FrozenDict
+
+# Fixme: update docstrings
 
 def tree_scalar_mul(scalar, tree_x):
     """Compute scalar * tree_x."""
@@ -130,32 +131,33 @@ def _root_vjp(
     """
     free_dofs_flat = None
     if free_dofs is not None:
-        free_dofs_flat = utility.dict_flatten(free_dofs)
+        dirichlet_dofs_flat = utility.dict_flatten(args[0]["dirichlet dofs"])
+        free_dofs_flat = ~dirichlet_dofs_flat
 
     @jax.custom_vjp
-    def linear_solver_fun_vjp(A, b):
-        return solve_fun(A, b)  # Here can be an external callback
+    def linear_solver_fun_vjp(A, b, free_dofs_flat):
+        return solve_fun(A, b, free_dofs_flat)  # Here can be an external callback
 
-    def linear_solver_fun_vjp_fwd(A, b):
-        result = linear_solver_fun_vjp(A, b)
-        return result, (A, b, result)
+    def linear_solver_fun_vjp_fwd(A, b, free_dofs_flat):
+        result = linear_solver_fun_vjp(A, b, free_dofs_flat)
+        return result, (A, b, free_dofs_flat, result)
 
     def linear_solver_fun_vjp_bwd(res, g):
-        A, b, result = res
+        A, b, free_dofs_flat, result = res
 
         # Ensure g matches the size of the expected gradient
         if free_dofs is not None:
 
             # Sparse outer product
-            result_dot = linear_solver_fun_vjp(A.T, g)
+            result_dot = linear_solver_fun_vjp(A.T, g, free_dofs_flat)
 
             size = A.shape[0]
-            empty = jnp.zeros((size,), dtype=jnp.float64)
+            empty = jnp.zeros((size,), dtype=float)
             result_tmp = utility.mask_op(empty, free_dofs_flat, result, "set")
             result_dot_tmp = utility.mask_op(empty, free_dofs_flat, result_dot, "set")
 
             # # Use the sparsity structure of A to build Fx_dot
-            indices = jnp.asarray(A.indices, dtype=jnp.int64)
+            indices = jnp.asarray(A.indices, dtype=int)
             data = -result_dot_tmp[indices[:, 0]] * result_tmp[indices[:, 1]]
 
             # Construct Fx_dot with the same sparsity pattern
@@ -163,10 +165,10 @@ def _root_vjp(
             Fy_dot = result_dot
         else:
             # Sparse outer product
-            result_dot = linear_solver_fun_vjp(A.T, g)
+            result_dot = linear_solver_fun_vjp(A.T, g, None)
 
             # Use the sparsity structure of A to build Fx_dot
-            indices = jnp.asarray(A.indices, dtype=jnp.int64)
+            indices = jnp.asarray(A.indices, dtype=int)
             data = -result_dot[indices[:, 0]] * result[indices[:, 1]]
 
             # Construct Fx_dot with the same sparsity pattern
@@ -174,7 +176,7 @@ def _root_vjp(
             Fy_dot = result_dot
 
         # Correctly return the shapes that match the input arguments
-        return (Fx_dot, Fy_dot)
+        return (Fx_dot, Fy_dot, None)
 
     linear_solver_fun_vjp.defvjp(linear_solver_fun_vjp_fwd, linear_solver_fun_vjp_bwd)
     diffable_solve_fun = linear_solver_fun_vjp
@@ -187,13 +189,10 @@ def _root_vjp(
     v = tree_scalar_mul(-1, cotangent)
 
[…]
