+++ autopdex/solver.py
@@ -37,39 +37,6 @@ from autopdex import assembler, implicit_diff, utility, implicit_diff_v2
 
 
 
-def custom_root_decorator(residual_fun, tangent_fun, lin_solve_fun, free_dofs=None, dirichlet_conditions=None, has_aux=False):
-    def decorator(root_solve_fun):
-        def wrapped(x):
-            # def tangent_solve_fun(g, y):
-            #     J = tangent_fun(x)
-            #     return lin_solve_fun(J, y)
-
-            def tangent_solve_fun(g, y):
-                J = tangent_fun(x)
-                solve = lambda matvec, x: lin_solve_fun(J, x)
-                transpose_solve = lambda vecmat, x: lin_solve_fun(J.T, x)
-                matvec = lambda x: utility.reshape_as(J @ utility.dict_flatten(x), x)
-                return jax.lax.custom_linear_solve(matvec, y, solve, transpose_solve)
-
-            solution = implicit_diff_v2.custom_root(
-                f=residual_fun,
-                initial_guess=x,
-                solve=lambda f_func, x0: root_solve_fun(x0),
-                tangent_solve=tangent_solve_fun,
-                free_dofs=free_dofs,
-                dirichlet_conditions=dirichlet_conditions,
-                has_aux=False
-            )
-            return solution
-            # if has_aux:
-            #     return solution, aux
-            # else:
-            #     return solution
-
-        return wrapped
-    return decorator
-
-
 ### Solvers as specified by the static_settings and settings
 @utility.jit_with_docstring(static_argnames=["static_settings", "**kwargs"])
 def solver(dofs, settings, static_settings, **kwargs):
@@ -308,21 +275,20 @@ def adaptive_load_stepping(
         #     lin_solve_fun, jnp.zeros(rhs.shape, rhs.dtype), mat, rhs, free_dofs_flat
         # )
 
-        # def lin_solve_callback_fun(mat, rhs):
-        #     rhs_flat = utility.dict_flatten(rhs)
-        #     # sol = jax.pure_callback(lin_solve_fun, jnp.zeros(rhs_flat.shape, rhs_flat.dtype), mat, rhs_flat, free_dofs_flat)
-        #     sol = jax.pure_callback(lin_solve_fun, jnp.zeros(rhs_flat.shape, rhs_flat.dtype), mat, rhs_flat, None)
-        #     return utility.reshape_as(sol, rhs)
-
-        def lin_solve_callback_fun(mat, rhs):
+        def lin_solve_callback_fun(mat, rhs, free_dofs_flat):
             rhs_flat = utility.dict_flatten(rhs)
-            # linear_solver = lambda mat, rhs: jax.pure_callback(lin_solve_fun, jnp.zeros(rhs_flat.shape, rhs_flat.dtype), mat, rhs_flat, None)
-            linear_solver = lambda mat, rhs: jax.pure_callback(lin_solve_fun, jnp.zeros(rhs_flat.shape, rhs_flat.dtype), mat, rhs_flat, free_dofs_flat)
-            solve = lambda matvec, x: linear_solver(mat, x)
-            transpose_solve = lambda vecmat, x: linear_solver(mat.T, x)
-            matvec = lambda x: utility.reshape_as(mat @ utility.dict_flatten(x), x)
-            sol = jax.lax.custom_linear_solve(matvec, rhs_flat, solve, transpose_solve)
+            sol = jax.pure_callback(lin_solve_fun, jnp.zeros(rhs_flat.shape, rhs_flat.dtype), mat, rhs_flat, free_dofs_flat)
             return utility.reshape_as(sol, rhs)
+
+        # def lin_solve_callback_fun(mat, rhs, free_dofs_flat):
+        #     rhs_flat = utility.dict_flatten(rhs)
[…]
