### 2024-12-09 a983587
integartion of changes
- added warnings regarding 'dirichlet dofs' and 'connectivity' if not shifted to settings
- updated examples accordingly
