+++ tests/test_mls_direct.py
@@ -101,9 +101,9 @@ def test_example_poisson():
     # 'hvp type': 'fwdrev',
 
     'verbose': 2,
-    'connectivity': (utility.jnp_to_tuple(neighbor_list),),
     })
     settings = {
+    'connectivity': (neighbor_list,),
     'node coordinates': x_nodes,
     'beta': (3,),
     'integration coordinates': (x_int,),
