+++ examples/cooks_membrane/quadrilaterals_p_refinement.py
@@ -94,11 +94,11 @@ for order in range(1,10):
     'solver backend': 'pardiso',
     'solver': 'lu',
     'verbose': 0,
-    'dirichlet dofs': utility.jnp_to_tuple(dirichlet_dofs),
-    'connectivity': (utility.jnp_to_tuple(elements),utility.jnp_to_tuple(neumann_elements)),
   })
 
   settings = {
+    'dirichlet dofs': dirichlet_dofs,
+    'connectivity': (elements, neumann_elements),
     'load multiplier': q_0,
     'node coordinates': x_nodes,
     'dirichlet conditions': dirichlet_conditions,
