+++ examples/laplace/fos_coupled_neumann_boundary.py
@@ -117,9 +117,9 @@ static_settings = flax.core.FrozenDict({
   'solver backend': 'pardiso',
   'solver': 'lu',
   'verbose': 1,
-  'connectivity': (utility.jnp_to_tuple(neighbor_list),),
 })
 settings = {
+  'connectivity': (neighbor_list,),
   'beta': (3,),
   'node coordinates': x_nodes,
   'integration coordinates': (x_int,),
