+++ examples/heat_conduction/space_time_fos_dirichlet_neumann_robin.py
@@ -106,9 +106,9 @@ static_settings = flax.core.FrozenDict({
   'solver backend': 'pardiso',
   'solver': 'lu',
   'verbose': 1,
-  'connectivity': (utility.jnp_to_tuple(neighbor_list),),
 })
 settings = {
+  'connectivity': (neighbor_list,),
   'beta': (3,),
   'node coordinates': x_nodes,
   'integration coordinates': (x_int,),

[…]
