+++ tests/test_user_elem_impl_diff_and_adaptive_load_step.py
@@ -18,7 +18,6 @@ def test_implicit_diff():
     from jax import config
     import jax.numpy as jnp
     import flax
-    import equinox
 
     from autopdex import seeder, geometry, solver, utility, models, spaces
 
