### 2025-01-14 #7300 @danielskatz
👋 @david-zwicker - you were previously willing to review this submission, but later than we wanted. However, we've lost a reviewer, so I wanted to check if you would be able to do the review now.
