### 2025-01-18 #7300 @david-zwicker
I had a brief look at the package, but since I never worked with the finite element method, I'm not sure I'm suited as a referee. I could have a look at general aspects of the Python package, but I won't be able to judge whether the API makes sense, whether typical use cases are covered, and how this package differs from others. I'm thus afraid a potential review would be borderline useless.
