### 2025-01-20 #7300 @danielskatz
@janvorisek - thanks.  I will add you now, and then you should follow the instructions in https://github.com/openjournals/joss-reviews/issues/7300#issuecomment-2380045610
