### 2025-02-17 #7300 @danielskatz
👋 @janvorisek - please provide any comments you have on the software & paper, and also please confirm that you think that this should be published now.
