### 2025-03-23 #7300 @janvorisek
@danielskatz @BodeTobias

Apologies for the delayed response. Below are my comments:

I have no significant concerns regarding the software or its implementation. From a FEA developer's perspective, it is well-documented, and the code is thoughtful and well-structured. The project is clearly well-designed, and the choice of tools and libraries is appropriate. I have also successfully run and tweaked around the provided examples.

I also went through the reviewer checklist again and can confirm that the project is suitable to be accepted.

## Performance Considerations

While formal benchmarking does not appear to have been conducted, it would be helpful for users to understand how **AutoPDEx** compares to other established solvers like **FreeFEM** or **MATLAB** in terms of performance.

## Documentation Feedback

A few suggestions to improve the documentation:

1. **Untranslated Sentences**: Some sentences in German are present on the [geometry page](https://bodetobias.github.io/AutoPDEx/geometry.html#geometry). These should be translated into English for clarity.

2. **Plotter Examples**: The [Plotter page](https://bodetobias.github.io/AutoPDEx/plotter.html) could benefit from example usage. Links to existing examples or sample code would help users better understand how to use the plotting functionality.

3. **Mesher Section and GMSH**: The documentation mentions a "Mesher" section, but **GMSH** appears to be used for meshing tasks. It would be helpful to mention **GMSH** in the mesher section and provide relevant examples or resources for clarity. May be it's a work in progress?

## Conclusion

These suggestions are intended to improve the usability and clarity of the software but are not critical. The documentation is strong, and with a few enhancements, it could possibly attract a wider audience.

I have no more comments at this stage. Thank you for the opportunity to review this project.
