+++ autopdex/dae.py
@@ -1,0 +1,2211 @@
+# dae.py
+# Copyright (C) 2025 Tobias Bode
+#
+# This program is free software: you can redistribute it and/or modify
+# it under the terms of the GNU Affero General Public License as published by
+# the Free Software Foundation, either version 3 of the License, or
+# (at your option) any later version.
+#
+# This program is distributed in the hope that it will be useful,
+# but WITHOUT ANY WARRANTY; without even the implied warranty of
+# MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+# GNU Affero General Public License for more details.
+
+"""Module for solving differential algebraic systems and transient PDEs.
+
+Das 'dae'-Modul beinhaltet Zeitintegrationsverfahren (Runge-Kutta und Mehrschrittverfahren) zum Lösen von gekoppelten gewöhnlichen Differentialgleichungen. Die Differentialgleichungen müssen in impliziter Form angegeben werden. Grundsätzlich kann man verschiedene Integratoren für verschiedene Felder wählen, die Anzahl und Position der Stufen müssen jedoch zueinander passen, da aktuell nur eine vollständig monolytische Lösung implementiert ist. Außerdem können Differential-Algebraische Systeme gelöst werden. Bei Vorliegen algebraischer Gleichungen sollten jedoch nur Integratoren mit ausschließlich impliziten Stufen verwendet werden. 
+
+Des Weiteren ist über das Keywort 'call pde' die Lösung von zeitabhängigen PDEs möglich (aktuell nur Zeitintegratoren mit einer Stage). Die Funktionsweise ist derart, dass die Knotenfreiheitsgrade über die gewählten Zeit-Diskretisierungsregeln als Funktionen der Zeit mit diskreten Werten und Ableitungen ausgestattet werden. Die räumliche Diskretisierung kann über die gleichen Ansätz (z.B. isoparametrische Finite Elemente) erfolgen, wie bei zeitunabhängigen Problemen. Im Falle von backward Euler und räumlichen FE-Ansatzfunktionen sähe das z.B. so aus:
+
+:math:`\Theta = \sum_{I=1}^{N} \Theta_I\left(t\right) \N_I\left(\boldsymbol{x}\right)`,
+
+wobei :math:`N_I\left(\boldsymbol{x}\right)` die FE-Ansatzfunktionen sind und :math:`\Theta_I\left(t\right)` die Knotenfreiheitsgrade sind. Die gegebene PDE wird dann im Schritt :math:`t_{n+1}` gelöst, wobei die Ableitung der Knotenfreiheitsgrade nach der Zeit durch eine custom_jvp-Regel mit :math:`\frac{\partial\Theta_I}{\partial t}\vert_{t_{n+1}} = \frac{\Theta_I\left(t_{n+1}\right) - \Theta_I\left(t_n\right)}{\Delta t}` ausgestattet wird.
+
+WARNING: This module is currently under development, experimental and has some undocumented limitations.
+"""
+
+# FIXME: docstrings for the integrators, controllers and plotting policies and TimeSteppingManager
+# FIXME: make sure, the integrator supports changing the step size if neccessarry! e.g. adams moulton: root iteration controler not possible...
+
+
+# TODO: generate tests from examples
+# TODO: translate all comments to english
+# TODO: staggered policies, explicit diagonal modes
+# TODO: add information about algebraic equations and add different treatements, e.g. projection for explicit modes
+
+
+from typing import Any
+from abc import ABC, abstractmethod
+from dataclasses import dataclass
+from math import isclose
+
+import numpy as np
+from scipy.sparse import csr_matrix
+from scipy.sparse.csgraph import connected_components
+
+import jax
+import jax.numpy as jnp
+from jax import tree_util, custom_jvp
[…]
