+++ docs/dae.rst
@@ -1,0 +1,17 @@
+DAE
+======
+
+.. automodule:: autopdex.dae
+    :no-index:
+
+.. autosummary::
+   :toctree: _autosummary
+
+   TimeSteppingManager
+
+   StepSizeController
+
+   SavePolicy
+
+   TimeIntegrator
+
