+++ docs/mesher.rst
@@ -8,4 +8,4 @@ Mesher
    :toctree: _autosummary
 
    structured_mesh
-
+   elevate_mesh_order
