+++ autopdex/dae.py
@@ -11,23 +11,9 @@
 # MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
 # GNU Affero General Public License for more details.
 
-"""Module for solving differential algebraic systems and transient PDEs.
+"""Module for solving differential algebraic systems and transient PDEs."""
 
-Das 'dae'-Modul beinhaltet Zeitintegrationsverfahren (Runge-Kutta und Mehrschrittverfahren) zum Lösen von gekoppelten gewöhnlichen Differentialgleichungen. Die Differentialgleichungen müssen in impliziter Form angegeben werden. Grundsätzlich kann man verschiedene Integratoren für verschiedene Felder wählen, die Anzahl und Position der Stufen müssen jedoch zueinander passen, da aktuell nur eine vollständig monolytische Lösung implementiert ist. Außerdem können Differential-Algebraische Systeme gelöst werden. Bei Vorliegen algebraischer Gleichungen sollten jedoch nur Integratoren mit ausschließlich impliziten Stufen verwendet werden. 
-
-Des Weiteren ist über das Keywort 'call pde' die Lösung von zeitabhängigen PDEs möglich (aktuell nur Zeitintegratoren mit einer Stage). Die Funktionsweise ist derart, dass die Knotenfreiheitsgrade über die gewählten Zeit-Diskretisierungsregeln als Funktionen der Zeit mit diskreten Werten und Ableitungen ausgestattet werden. Die räumliche Diskretisierung kann über die gleichen Ansätz (z.B. isoparametrische Finite Elemente) erfolgen, wie bei zeitunabhängigen Problemen. Im Falle von backward Euler und räumlichen FE-Ansatzfunktionen sähe das z.B. so aus:
-
-:math:`\Theta = \sum_{I=1}^{N} \Theta_I\left(t\right) \N_I\left(\boldsymbol{x}\right)`,
-
-wobei :math:`N_I\left(\boldsymbol{x}\right)` die FE-Ansatzfunktionen sind und :math:`\Theta_I\left(t\right)` die Knotenfreiheitsgrade sind. Die gegebene PDE wird dann im Schritt :math:`t_{n+1}` gelöst, wobei die Ableitung der Knotenfreiheitsgrade nach der Zeit durch eine custom_jvp-Regel mit :math:`\frac{\partial\Theta_I}{\partial t}\vert_{t_{n+1}} = \frac{\Theta_I\left(t_{n+1}\right) - \Theta_I\left(t_n\right)}{\Delta t}` ausgestattet wird.
-
-WARNING: This module is currently under development, experimental and has some undocumented limitations.
-"""
-
-# FIXME: docstrings for the integrators, controllers and plotting policies and TimeSteppingManager
-# FIXME: make sure, the integrator supports changing the step size if neccessarry! e.g. adams moulton: root iteration controler not possible...
-
-
+# TODO: make sure, the integrator supports changing the step size if neccessarry! e.g. adams moulton: root iteration controler not possible...
 # TODO: generate tests from examples
 # TODO: translate all comments to english
 # TODO: staggered policies, explicit diagonal modes
@@ -78,11 +64,12 @@ def discrete_value_with_derivatives(t, q, q_derivs):
   derivative `t_dot` to compute the overall derivative contribution.
 
   It is used to construct a differentiable q_fun based on a value and its derivative defined by an integration rule, e.g.:
-  ```
-  def diffable_q_fun(t): # q_ts is a tuple of (q, q_dot, q_ddot, ...) coming from the integrator
-    return {key: discrete_value_with_derivatives(t, q_ts[key][0], q_ts[key][1:]) for key in template.keys()}
-  ```
 
+    .. code-block:: python
+    
[…]
