+++ docs/example_notebooks.rst
@@ -7,3 +7,6 @@ Example notebooks
 
    notebooks/quickstart
    notebooks/mixed_problem
+   notebooks/ode_lotka_volterra
+   notebooks/dae_robertson
+   notebooks/pdae_navier_stokes_and_energy_balance
