+++ docs/mesher.rst
@@ -2,7 +2,6 @@ Mesher
 ======
 
 .. automodule:: autopdex.mesher
-    :no-index:
 
 .. autosummary::
    :toctree: _autosummary
