### 2025-04-08 53bd4a9
Added Manifest.in
