+++ MANIFEST.in
@@ -1,0 +1,3 @@
+prune docs
+prune examples
+prune paper
