+++ pyproject.toml
@@ -8,12 +8,11 @@ version = "1.1.1"
 description = "An automized PDE solver for accelerators based on JAX"
 readme = "README.md"
 authors = [
-    { name = "Tobias Bode", email = "tobiasbode.94@gmail.com" }
+    { name = "Tobias Bode" }
 ]
-license = { text = "GNU Affero General Public License v3" }
+license = "AGPL-3.0-only"
 classifiers = [
     "Programming Language :: Python :: 3",
-    "License :: OSI Approved :: GNU Affero General Public License v3",
     "Operating System :: OS Independent"
 ]
 requires-python = ">=3.10"
@@ -30,7 +29,6 @@ dependencies = [
     "pyvista>=0.43.3",
     "scipy>=1.13.0",
     "scikit-optimize>=0.9.0",
-    "sparsejac>=0.1.3",
     "pytest>=8.3.2"
 ]
 
