### 2025-04-10 #7300 @BodeTobias
@janvorisek @danielskatz

Thanks a lot for reviewing the project! I have uploaded a new version of the package and partially integrated the suggestions.
From my point of view, the paper is now ready for publication. Below you find my answer to the feedback of @janvorisek

- Performance considerations: I haven't done detailed benchmarks so far, but I will keep that in mind. Overall, the performance depends very much on the problem, hardware and the chosen external linear solver. However, I have compared the Poisson problem from the GitHub README with Ferrite.jl and AceFEM and found the computational time to be in the same range.
- Untranslated sentences: Thanks for finding that. I have translated it to English.
- Plotter examples and mesher section: The functionality in this regard is still very limited and mostly for making the development and testing easier. I have added statements in the docs and recommend to use, e.g., GMSH for meshing and meshio for writing VTK files for ParaView. More convenience functions for easier pre- and postprocessing are also on my to-do list.
