### 2025-04-10 #7300 @danielskatz
I've suggested some changes in https://github.com/BodeTobias/AutoPDEx/pull/5 - please merge this, or let me know what you disagree with, then we can move on the next steps.
