+++ pyproject.toml
@@ -4,7 +4,7 @@ build-backend = "setuptools.build_meta"
 
 [project]
 name = "autopdex"
-version = "1.1.1"
+version = "1.1.4"
 description = "An automized PDE solver for accelerators based on JAX"
 readme = "README.md"
 authors = [
