### 2025-04-10 #7300 @BodeTobias
Ok, I have made a new tagged release and generated a DOI for it.

Version tag: v1.1.4
DOI: 10.5281/zenodo.15192270
