### 2025-04-11 #7300 @editorialbot
:wave: @openjournals/csism-eics, this paper is ready to be accepted and published.
