### 2025-04-11 #7300 @danielskatz
Congratulations to  @BodeTobias on your JOSS publication!!

And thanks to @celliern & @janvorisek for reviewing!
